basic client implementation
