+++ client/src/Client.ts
@@ -1,161 +1,63 @@
-import * as protocol_root from './proto';
-import Message = protocol_root.mozaic.protocol.Message;
-import { BotRunner, BotConfig } from "./BotRunner";
-import { Connection, Address } from "./Connection";
-import { Socket } from 'net';
-import { BufferWriter } from 'protobufjs';
-import { ClientLogger, Logger } from './Logger';
-import { TextDecoder } from 'text-encoding';
-import { ServerMessage, GameState, PlayerAction } from './PwTypes';
-import { RequestResolver } from './RequestResolver';
+import { Logger, Connection } from "./index";
+import { LogRecord } from "./PwTypes";
 import { SimpleEventDispatcher, ISimpleEvent } from 'ste-simple-events';
-import { SignalDispatcher, ISignal } from 'ste-signals';
-
-export interface ConnectionData {
-    token: Buffer,
-    address: Address,
-}
-
-export interface Request {
-    requestId: number;
-    data: Buffer;
-}
-
-// TODO: expand this
-enum ClientState {
-    CONNECTING,
-    CONNECTED,
-};
 
 export interface ClientParams {
-    token: Buffer,
-    address: Address,
-    number: number,
-    botConfig: BotConfig,
-    logger: Logger,
+    host: string;
+    port: number;
+    token: Uint8Array;
+    logger: Logger;
+}
+
+export interface ClientProps {
+    clientId: number;
+    connection: Connection;
+    logger: Logger;
 }
 
 export class Client {
-    readonly connection: Connection;
-    readonly address: Address;
-    readonly botRunner: BotRunner;
-    readonly logger: ClientLogger;
-    private resolver: RequestResolver;
+    clientId: number;
+    connection: Connection;
+    logger: Logger;
 
-    private turnNum: 0;
-    private state: ClientState;
-
-    private _onExit = new SignalDispatcher();
-    private _onError = new SimpleEventDispatcher<Error>();
-
-    constructor(params: ClientParams) {
-        this.connection = new Connection(params.token);
-
-        this.handleMessage = this.handleMessage.bind(this);
-        this.resolver = new RequestResolver(this.connection, this.handleMessage);
-
-
-        this.address = params.address;
-        this.botRunner = new BotRunner(params.botConfig);
-        this.logger = new ClientLogger(params.logger, params.number);
-        this.state = ClientState.CONNECTED;
-        this.turnNum = 0;
-
-        this.botRunner.onError.subscribe((err) => {
-            this._onError.dispatch(err);
-        });
-
-        this.connection.onError.subscribe((err) => {
-            this._onError.dispatch(err);
-        });
-
-        this.connection.onClose.subscribe(() => {
-            this.botRunner.killBot();
-            this._onExit.dispatch();
-        });
-    }
-
-    public run() {
-        this.connection.connect(this.address.host, this.address.port);
-        this.botRunner.run();
-    }
-
-    public get onExit() {
-        return this._onExit.asEvent();
-    }
-
-    public get onError() {
-        return this._onError.asEvent();
-    }
-
-    private handleMessage(data: Uint8Array): void | Promise<Uint8Array> {
-        switch (this.state) {
-            case ClientState.CONNECTING: {
-                return this.handleConnectingMessage(data);
-            }
-            case ClientState.CONNECTED: {
-                return this.handleGameMessage(data);
-            }
-        }
-    }
-
-    private handleConnectingMessage(data: Uint8Array): Promise<Uint8Array> {
-        this.state = ClientState.CONNECTED;
-        return Promise.resolve(data);
-    }
-
-    private handleGameMessage(data: Uint8Array): void | Promise<Uint8Array>
-    {
-        const text = new TextDecoder('utf-8').decode(data);
-        let serverMessage: ServerMessage = JSON.parse(text);
-
-        switch (serverMessage.type) {
-            case 'game_state': {
-                return this.handleGameState(serverMessage.content);
-            }
-            case 'player_action': {
-                return this.handlePlayerAction(serverMessage.content);
-            }
-            case 'final_state': {
-                return this.handleFinalState(serverMessage.content);
-            }
-        }
-    }
-
-    private handleGameState(state: GameState): Promise<Uint8Array> {
-        this.logState(state);
-
+    public static connect(params: ClientParams): Promise<Client> {
         return new Promise((resolve, reject) => {
-            const request = JSON.stringify(state);
-            this.botRunner.request(request, (response) => {
-                this.logger.log({
-                    "type": "command",
-                    "content": response,
+            const connection = new Connection(params.token);
+            connection.onConnect.one((clientId) => {
+                const client = new Client({
+                    clientId,
+                    connection,
+                    logger: params.logger,
                 });
-                const buf = Buffer.from(response, 'utf-8');
-                resolve(buf);
+                resolve(client);
             });
+            connection.onError.one((err) => {
+                reject(err);
+            });
+
+            connection.connect(params.host, params.port);
         });
     }
 
-    private handleFinalState(state: GameState) {
-        this.logState(state);
+    constructor(props: ClientProps) {
+        this.clientId = props.clientId;
+        this.connection = props.connection;
+        this.logger = props.logger;
     }
 
-    private logState(state: GameState) {
-        this.turnNum += 1;
-
+    public log(record: LogRecord) {
         this.logger.log({
-            "type": 'step',
-            "turn_number": this.turnNum,
-            "state": state,
+            type: "player_entry",
+            player: this.clientId,
+            record,
         });
     }
 
-    private handlePlayerAction(action: PlayerAction) {
-        this.logger.log({
-            "type": 'player_action',
-            "action": action,
-        })
-    }   
+    public send_message(message: Uint8Array) {
+        this.connection.send(message);
+    }
+
+    public get onMessage() {
+        return this.connection.onMessage;
+    }
 }
